feat: open event form modal from base screen
